Adding Makefile for easier iteration/development
This commit adds a Makefile with a few helpful targets to allow for
somewhat easier iteration while doing development. This helps the
developer not have to remember all the incantations for Docker and will
properly clean the workspace before building an image (because otherwise
all previous build artifacts are hoisted into the container).

Includes a little start script for the container with env var checking.

README is also updated appropriately.